Adds a python script for chrooting in to the rootfs images
This should work for all distros since we do the binary patching of
FEXInterpreter at all times for this script. Python is a lot easier to
work with for all the considerations we need to make. A big improvement
is just checking that all applications are installed before actually
executing. Plus more error checking if mount points are mangled. Overall
an improvement, and will allow more improvements going forward.

Once the rootfs images are switched over to this script then the bash
scripts will go away.